[BFCL] Reorganized All `constant.py` Files to a `constants` Folder (#944)
Reorganized all constants in the following files in a separate folder
`./berkeley-function-call-leaderboard/bcfl/constants` for clarity.
- `berkeley-function-call-leaderboard/bfcl/constant.py` ->
`category_mapping.py` & `eval_config.py`
- `berkeley-function-call-leaderboard/bfcl/eval_checker/constant.py` ->
`eval_config.py` & `column_headers.py`
- `berkeley-function-call-leaderboard/bfcl/model_handler/constant.py` ->
`type_mappings.py` & `eval_config.py`(`UNDERSCORE_TO_DOT` specifically)
-
`berkeley-function-call-leaderboard/bfcl/model_handler/local_inference/constant.py`
-> `eval_config.py`

---------
