+++ berkeley-function-call-leaderboard/bfcl/constants/category_mapping.py
@@ -1,16 +1,3 @@
-from pathlib import Path
-
-# NOTE: These paths are relative to the `bfcl` directory where this script is located.
-RESULT_PATH = "../result/"
-PROMPT_PATH = "../data/"
-MULTI_TURN_FUNC_DOC_PATH = "../data/multi_turn_func_doc/"
-POSSIBLE_ANSWER_PATH = "../data/possible_answer/"
-SCORE_PATH = "../score/"
-DOTENV_PATH = "../.env"
-UTILS_PATH = "../utils/"
-PROJECT_ROOT = "../"
-TEST_IDS_TO_GENERATE_PATH = "../test_case_ids_to_generate.json"
-
 VERSION_PREFIX = "BFCL_v3"
 
 # These are in the PROMPT_PATH
@@ -190,19 +177,3 @@ MULTI_TURN_FUNC_DOC_FILE_MAPPING = {
     "TravelAPI": "travel_booking.json",
     "VehicleControlAPI": "vehicle_control.json",
 }
-
-
-# Construct the full path to use by other scripts
-script_dir = Path(__file__).parent
-RESULT_PATH = (script_dir / RESULT_PATH).resolve()
-PROMPT_PATH = (script_dir / PROMPT_PATH).resolve()
-MULTI_TURN_FUNC_DOC_PATH = (script_dir / MULTI_TURN_FUNC_DOC_PATH).resolve()
-POSSIBLE_ANSWER_PATH = (script_dir / POSSIBLE_ANSWER_PATH).resolve()
-SCORE_PATH = (script_dir / SCORE_PATH).resolve()
-DOTENV_PATH = (script_dir / DOTENV_PATH).resolve()
-UTILS_PATH = (script_dir / UTILS_PATH).resolve()
-PROJECT_ROOT = (script_dir / PROJECT_ROOT).resolve()
-TEST_IDS_TO_GENERATE_PATH = (script_dir / TEST_IDS_TO_GENERATE_PATH).resolve()
-
-RESULT_PATH.mkdir(parents=True, exist_ok=True)
-SCORE_PATH.mkdir(parents=True, exist_ok=True)
