+++ berkeley-function-call-leaderboard/bfcl/constants/type_mappings.py
@@ -1,0 +1,89 @@
+GORILLA_TO_OPENAPI = {
+    "integer": "integer",
+    "number": "number",
+    "float": "number",
+    "string": "string",
+    "boolean": "boolean",
+    "bool": "boolean",
+    "array": "array",
+    "list": "array",
+    "dict": "object",
+    "object": "object",
+    "tuple": "array",
+    "any": "string",
+    "byte": "integer",
+    "short": "integer",
+    "long": "integer",
+    "double": "number",
+    "char": "string",
+    "ArrayList": "array",
+    "Array": "array",
+    "HashMap": "object",
+    "Hashtable": "object",
+    "Queue": "array",
+    "Stack": "array",
+    "Any": "string",
+    "String": "string",
+    "Bigint": "integer",
+}
+
+GORILLA_TO_PYTHON = {
+    "integer": "int",
+    "number": "float",
+    "float": "float",
+    "string": "str",
+    "boolean": "bool",
+    "bool": "bool",
+    "array": "list",
+    "list": "list",
+    "dict": "dict",
+    "object": "dict",
+    "tuple": "tuple",
+    "any": "str",
+    "byte": "int",
+    "short": "int",
+    "long": "int",
+    "double": "float",
+    "char": "str",
+    "ArrayList": "list",
+    "Array": "list",
+    "HashMap": "dict",
+    "Hashtable": "dict",
+    "Queue": "list",
+    "Stack": "list",
+    "Any": "str",
+    "String": "str",
+    "Bigint": "int",
+}
+
+
+JAVA_TYPE_CONVERSION = {
+    "byte": int,
+    "short": int,
+    "integer": int,
+    "float": float,
+    "double": float,
+    "long": int,
+    "boolean": bool,
+    "char": str,
+    "Array": list,
+    "ArrayList": list,
+    "Set": set,
+    "HashMap": dict,
+    "Hashtable": dict,
+    "Queue": list,  # this can be `queue.Queue` as well, for simplicity we check with list
+    "Stack": list,
+    "String": str,
+    "any": str,
+}
+
+JS_TYPE_CONVERSION = {
+    "String": str,
+    "integer": int,
+    "float": float,
+    "Bigint": int,
+    "Boolean": bool,
+    "dict": dict,
+    "array": list,
+    "any": str,
+}
