+++ berkeley-function-call-leaderboard/bfcl/eval_checker/ast_eval/type_convertor/java_type_converter.py
@@ -1,6 +1,6 @@
 import re
 from typing import List, Dict, Union
-from bfcl.model_handler.constant import JAVA_TYPE_CONVERSION
+from bfcl.constants.type_mappings import JAVA_TYPE_CONVERSION
 
 
 def java_type_converter(value, expected_type, nested_type=None):
