+++ berkeley-function-call-leaderboard/bfcl/model_handler/api_inference/mistral.py
@@ -3,7 +3,7 @@ import os
 import time
 
 from bfcl.model_handler.base_handler import BaseHandler
-from bfcl.model_handler.constant import GORILLA_TO_OPENAPI
+from bfcl.constants.type_mappings import GORILLA_TO_OPENAPI
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.utils import (
     ast_parse,
