+++ berkeley-function-call-leaderboard/bfcl/model_handler/api_inference/nova.py
@@ -3,7 +3,7 @@ import time
 
 import boto3
 from bfcl.model_handler.base_handler import BaseHandler
-from bfcl.model_handler.constant import GORILLA_TO_OPENAPI
+from bfcl.constants.type_mappings import GORILLA_TO_OPENAPI
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.utils import (
     combine_consecutive_user_prompts,
