+++ berkeley-function-call-leaderboard/bfcl/model_handler/base_handler.py
@@ -2,16 +2,17 @@ import json
 import time
 from copy import deepcopy
 
-from bfcl.constant import RESULT_PATH, VERSION_PREFIX
+from bfcl.constants.category_mapping import VERSION_PREFIX
+from bfcl.constants.default_prompts import (
+    DEFAULT_USER_PROMPT_FOR_ADDITIONAL_FUNCTION_FC,
+    DEFAULT_USER_PROMPT_FOR_ADDITIONAL_FUNCTION_PROMPTING,
+    MAXIMUM_STEP_LIMIT,
+)
+from bfcl.constants.eval_config import RESULT_PATH
 from bfcl.eval_checker.multi_turn_eval.multi_turn_utils import (
     STATELESS_CLASSES,
     execute_multi_turn_func_call,
     is_empty_execute_response,
-)
-from bfcl.model_handler.constant import (
-    DEFAULT_USER_PROMPT_FOR_ADDITIONAL_FUNCTION_FC,
-    DEFAULT_USER_PROMPT_FOR_ADDITIONAL_FUNCTION_PROMPTING,
-    MAXIMUM_STEP_LIMIT,
 )
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.utils import load_file, make_json_serializable, sort_key
