+++ berkeley-function-call-leaderboard/bfcl/model_handler/local_inference/base_oss_handler.py
@@ -5,10 +5,12 @@ import os
 from concurrent.futures import ThreadPoolExecutor
 
 import requests
-from bfcl.constant import RESULT_PATH
+from bfcl.constants.eval_config import (
+    RESULT_PATH,
+    VLLM_PORT,
+)
 from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.model_style import ModelStyle
-from bfcl.model_handler.local_inference.constant import VLLM_PORT
 from bfcl.model_handler.utils import (
     default_decode_ast_prompting,
     default_decode_execute_prompting,
