--- berkeley-function-call-leaderboard/bfcl/model_handler/local_inference/constant.py
@@ -1,1 +1,0 @@
-VLLM_PORT = 1053
