+++ berkeley-function-call-leaderboard/bfcl/model_handler/utils.py
@@ -7,7 +7,8 @@ import re
 from functools import reduce
 from typing import Callable, List, Optional, Type, Union
 
-from bfcl.model_handler.constant import DEFAULT_SYSTEM_PROMPT, GORILLA_TO_OPENAPI
+from bfcl.constants.default_prompts import DEFAULT_SYSTEM_PROMPT
+from bfcl.constants.type_mappings import GORILLA_TO_OPENAPI
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.parser.java_parser import parse_java_function_call
 from bfcl.model_handler.parser.js_parser import parse_javascript_function_call
