+++ berkeley-function-call-leaderboard/openfunctions_evaluation.py
@@ -1,5 +1,5 @@
 from bfcl._llm_response_generation import get_args, main
-from bfcl.constant import DOTENV_PATH
+from bfcl.constants.eval_config import DOTENV_PATH
 from dotenv import load_dotenv
 
 # Note: This file is still kept for compatibility with the old structure of the codebase.
